backend/requirements.txt
# Core Framework
eel==0.16.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
sqlite-utils==3.35.2

# Data Analysis
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.7
pypdf==3.17.4

# Validation
pydantic==2.5.2
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1